Upgraded filter results and date data
There were a few issues with the air dates which I fixed. Now all data should be good, but there might be some issues that I missed. Dateparser is a MESS. Do NOT try to understand it. I just fixed issues as they came up, I did not want to refactor that code.